pyproject.toml
[project]
name = "ready-tensor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
]
